customer-service/requirements.txt
Flask==3.0.3
mysql-connector-python==9.1.0
kafka-python==2.0.2
